server/requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.3
pydantic-settings==2.8.1
python-dotenv==1.1.0

# LangChain core
langchain==0.3.23
langchain-openai==0.3.13
langchain-core==0.3.52
langchain-community==0.3.21
langgraph==0.3.31

# Vector store
langchain_pinecone==0.2.5
pinecone-client==6.0.0

# LLM and search
openai==1.75.0
tavily-python==0.5.4

# New dependencies
python-docx>=0.8.11
PyPDF2>=3.0.0

# Other dependencies
python-multipart

# Twilio
twilio>=8.2.0
aiohttp>=3.8.0  # For async HTTP requests

